separate main logic from fuse ops handlers
